practice_mafia/requirements.txt
beautifulsoup4==4.12.2
google==3.0.0
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
soupsieve==2.4.1
